+++ internal/framework/services/objectstorage/resource_bucket_test.go
@@ -14,6 +14,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/acctest"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/identity"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/internal/framework/services/objectstorage"
 )
 
 func TestAccBucketResource(t *testing.T) {
@@ -35,6 +37,11 @@ func TestAccBucketResource(t *testing.T) {
 					resource.TestCheckResourceAttr(name, "region", "eu-central-3"),
 					resource.TestCheckResourceAttr(name, "id", rName),
 					resource.TestCheckResourceAttr(name, "object_lock_enabled", "false"),
+					acctest.TestAccCheckResourceInList(
+						objectstorage.NewBucketListResource(),
+						identity.WithIdentity(objectstorage.NewBucketResource)(),
+						rName, false,
+					),
 				),
 			},
 			{